refactor and fix: imports, redirects
