+++ src/customHooks/index.js
@@ -1,0 +1,1 @@
+export { useDocumentTitle } from "./useDocumentTitle";
