+++ src/pages/Home/Home.jsx
@@ -3,7 +3,7 @@ import { banner2 } from "../../data/image/productImages";
 import { categories } from "../../backend/db/categories";
 import { products } from "../../backend/db/products";
 import "./Home.css";
-import { useDocumentTitle } from "../../customHooks/useDocumentTitle";
+import { useDocumentTitle } from "../../customHooks";
 import { CategoryListing } from "../../pages";
 import { Link } from "react-router-dom";
